+++ front/src/types.ts
@@ -1,0 +1,6 @@
+type FAQ = {
+    question: string;
+    pageTitle: string;
+};
+
+export type { FAQ };
